Refactor streamlit app into pages
